refactor: update the files structure
