+++ src/commands/hash.js
@@ -1,0 +1,1 @@
+
