+++ src/commands/navwd.js
@@ -1,0 +1,1 @@
+
